Make cuda an extension
