+++ Project.toml
@@ -6,7 +6,6 @@ version = "0.13.2"
 [deps]
 Adapt = "79e6a3ab-5dfb-504d-930d-738a2a938a0e"
 Artifacts = "56f22d72-fd6d-98f1-02f0-08ddc0907c33"
-CUDA = "052768ef-5323-5732-b1bb-66c8b64840ba"
 ClimaComms = "3a4d1b5c-c61d-41fd-a00a-5873ba7a1b0d"
 DocStringExtensions = "ffbed154-4ef7-542d-bbb7-c09d3a79fcae"
 GaussQuadrature = "d54b0c1a-921d-58e0-8e36-89d8069c0969"
@@ -15,16 +14,18 @@ StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
 ClimaParams = "5c42b081-d73a-476f-9059-fd94b934656c"
 
 [weakdeps]
+CUDA = "052768ef-5323-5732-b1bb-66c8b64840ba"
 ClimaParams = "5c42b081-d73a-476f-9059-fd94b934656c"
 
 [extensions]
 CreateParametersExt = "ClimaParams"
+RRTMGPCUDAExt = "CUDA"
 
 [compat]
 Adapt = "3.3, 4"
 Artifacts = "1"
 CUDA = "4, 5"
-ClimaComms = "0.5.1"
+ClimaComms = "0.5.1, 0.6"
 DocStringExtensions = "0.8, 0.9"
 GaussQuadrature = "0.5"
 Random = "1"
